Implement ST-Link communication for probing attack analysis
There now exists the STLinkComm class, a slim wrapper around some pystlink library classes that allows us to
- get register values and names,
- get available register names, and,
- reset the device under test, optionally halting.

This class is mostly compatible with the OpenOCD class we worked with before. I implemented a RegisterReading dataclass and generified over
a _ReadingType to better reflect the flexibility that value_cast callback provides for us. The RegisterReading dataclass also clarifies
the interface for register values instead of relying on the existence of specific dictionary keys.

CAVEATS: As I have developed this at home, and not in the lab, I have not yet tested how STLink behaves when exposed to ChipSHOUTER, so this
implementation might break arbitrarily once exposed to EMFI pulses.